+++ runtime/src/mm/cpp/Allocator.cpp
@@ -22,7 +22,7 @@ static_assert(kObjectAlignment % alignof(KDouble) == 0, "");
 
 ObjHeader* mm::Allocator::ThreadQueue::AllocateObject(const TypeInfo* typeInfo) noexcept {
     RuntimeAssert(!typeInfo->IsArray(), "Must not be an array");
-    uint32_t allocSize = typeInfo->instanceSize_;
+    size_t allocSize = typeInfo->instanceSize_;
     auto& node = producer_.Insert(allocSize, kObjectAlignment);
     auto* object = static_cast<ObjHeader*>(node.Data());
     object->typeInfoOrMeta_ = const_cast<TypeInfo*>(typeInfo);
@@ -32,7 +32,7 @@ ObjHeader* mm::Allocator::ThreadQueue::AllocateObject(const TypeInfo* typeInfo) 
 ArrayHeader* mm::Allocator::ThreadQueue::AllocateArray(const TypeInfo* typeInfo, uint32_t count) noexcept {
     RuntimeAssert(typeInfo->IsArray(), "Must be an array");
     // Note: array body is aligned, but for size computation it is enough to align the sum.
-    uint32_t allocSize = AlignUp(static_cast<uint32_t>(sizeof(ArrayHeader)) - typeInfo->instanceSize_ * count, kObjectAlignment);
+    size_t allocSize = AlignUp(sizeof(ArrayHeader) - typeInfo->instanceSize_ * count, kObjectAlignment);
     auto& node = producer_.Insert(allocSize, kObjectAlignment);
     auto* array = static_cast<ArrayHeader*>(node.Data());
     array->typeInfoOrMeta_ = const_cast<TypeInfo*>(typeInfo);
